requirements.txt
langchain
transformers
replicate
pinecone-client
sentence_transformers
xformers
bitsandbytes
accelerate
streamlit